Organized classes into "Models" folder
